fix(earn): address review feedback
