+++ apps/sdp-web/src/app/dashboard/markets/earn/loading.tsx
@@ -1,5 +1,1 @@
-import { EarnProgramSkeleton } from "../markets-route-skeletons";
-
-export default function EarnLoading() {
-  return <EarnProgramSkeleton />;
-}
+export { default } from "../embedded-yield/loading";
