+++ apps/sdp-web/src/app/dashboard/markets/earn/page.tsx
@@ -1,16 +1,3 @@
-import { DASHBOARD_MARKETS_SUBNAV_HREFS } from "@/lib/dashboard-navigation-loading";
-import { EarnProgramWorkspace } from "./earn-program-workspace";
-import { loadEarnProviderAccess } from "./earn-provider-access.server";
+export { default } from "../embedded-yield/page";
 
-/** Provider access is organization-scoped and must be resolved per request. */
 export const dynamic = "force-dynamic";
-
-export default async function EarnPage() {
-  const providerAccess = await loadEarnProviderAccess();
-  return (
-    <EarnProgramWorkspace
-      builderHref={`${DASHBOARD_MARKETS_SUBNAV_HREFS.earnProgram}/button-builder`}
-      providerAccess={providerAccess}
-    />
-  );
-}
