+++ apps/sdp-web/src/app/dashboard/markets/embedded-yield/button-builder/loading.tsx
@@ -1,1 +1,5 @@
-export { default } from "../../earn/button-builder/loading";
+import { EarnProgramSkeleton } from "../../markets-route-skeletons";
+
+export default function EmbeddedYieldButtonBuilderLoading() {
+  return <EarnProgramSkeleton />;
+}
