+++ apps/sdp-web/src/app/dashboard/markets/embedded-yield/loading.tsx
@@ -1,1 +1,5 @@
-export { default } from "../earn/loading";
+import { EarnProgramSkeleton } from "../markets-route-skeletons";
+
+export default function EmbeddedYieldLoading() {
+  return <EarnProgramSkeleton />;
+}
