+++ apps/sdp-web/src/app/dashboard/markets/markets-route-loading.unit.test.tsx
@@ -1,20 +1,17 @@
 import { renderToStaticMarkup } from "react-dom/server";
 import { describe, expect, it } from "vitest";
-import EarnButtonBuilderLoading from "./earn/button-builder/loading";
-import EarnLoading from "./earn/loading";
 import EmbeddedYieldButtonBuilderLoading from "./embedded-yield/button-builder/loading";
 import EmbeddedYieldLoading from "./embedded-yield/loading";
 
 describe("Markets route loading states", () => {
-  it("preserves the Earn workspace skeleton on the canonical route", () => {
-    expect(renderToStaticMarkup(<EmbeddedYieldLoading />)).toBe(
-      renderToStaticMarkup(<EarnLoading />)
-    );
-  });
+  it.each([
+    ["workspace", EmbeddedYieldLoading],
+    ["button builder", EmbeddedYieldButtonBuilderLoading],
+  ])("renders the complete Embedded Yield %s skeleton", (_name, Loading) => {
+    const markup = renderToStaticMarkup(<Loading />);
 
-  it("preserves the Earn builder skeleton on the canonical route", () => {
-    expect(renderToStaticMarkup(<EmbeddedYieldButtonBuilderLoading />)).toBe(
-      renderToStaticMarkup(<EarnButtonBuilderLoading />)
-    );
+    expect(markup).toContain('aria-busy="true"');
+    expect(markup).toContain("max-w-7xl");
+    expect(markup.match(/animate-pulse/g)).toHaveLength(14);
   });
 });
