+++ apps/sdp-web/src/i18n/messages.ts
@@ -47,7 +47,7 @@ import viDashboardPrivateChannels from "../../messages/vi/dashboard-private-chan
 import viShared from "../../messages/vi/shared.json";
 import vi from "../../messages/vi.json";
 
-const enMessages = {
+export const englishSourceMessages = {
   ...en,
   ...dashboardApprovals,
   ...dashboardCustody,
@@ -60,7 +60,7 @@ const enMessages = {
   Shared: shared,
 };
 
-export type Messages = typeof enMessages;
+export type Messages = typeof englishSourceMessages;
 
 type LocalizedMessages<TValue> = {
   [TKey in keyof TValue]?: TValue[TKey] extends string ? string : LocalizedMessages<TValue[TKey]>;
@@ -73,29 +73,25 @@ export function mergeLocalizedMessages<TValue>(
   return mergeLocalizedValue(fallback, localized) as TValue;
 }
 
-const legacyEarnProductNamePattern = /\bEarn\b/g;
-
-function normalizeEmbeddedYieldBrand<TValue>(value: TValue): TValue {
-  // Localized catalogs land independently and can retain the previous product
-  // name. Keep the rendered brand canonical while their copy catches up.
-  if (typeof value === "string") {
-    return value.replace(legacyEarnProductNamePattern, "Embedded Yield") as TValue;
-  }
-  if (Array.isArray(value)) {
-    return value.map(normalizeEmbeddedYieldBrand) as TValue;
-  }
-  if (!value || typeof value !== "object") {
-    return value;
-  }
-
-  return Object.fromEntries(
-    Object.entries(value).map(([key, nested]) => [key, normalizeEmbeddedYieldBrand(nested)])
-  ) as TValue;
+export function mergeLocalizedMessagesWithEmbeddedYieldBrand<TValue>(
+  fallback: TValue,
+  localized: LocalizedMessages<TValue> | undefined
+): TValue {
+  // Temporary bridge: remove this variant after the translation release PR
+  // replaces legacy Earn brand references in every localized catalog.
+  return mergeLocalizedValue(fallback, localized, true) as TValue;
 }
 
-function mergeLocalizedValue(fallback: unknown, localized: unknown): unknown {
+function mergeLocalizedValue(
+  fallback: unknown,
+  localized: unknown,
+  preserveEmbeddedYieldBrand = false
+): unknown {
   if (typeof fallback === "string") {
-    return typeof localized === "string" ? localized : fallback;
+    const resolved = typeof localized === "string" ? localized : fallback;
+    return preserveEmbeddedYieldBrand && fallback.includes("Embedded Yield")
+      ? resolved.replace(/\bEarn\b/g, "Embedded Yield")
+      : resolved;
   }
   if (!fallback || typeof fallback !== "object" || Array.isArray(fallback)) {
     return fallback;
@@ -108,7 +104,7 @@ function mergeLocalizedValue(fallback: unknown, localized: unknown): unknown {
   return Object.fromEntries(
     Object.entries(fallback).map(([key, fallbackValue]) => [
       key,
-      mergeLocalizedValue(fallbackValue, localizedRecord[key]),
+      mergeLocalizedValue(fallbackValue, localizedRecord[key], preserveEmbeddedYieldBrand),
     ])
   );
 }
@@ -124,7 +120,7 @@ const esCatalog = {
   Shared: esShared,
 } satisfies LocalizedMessages<Messages>;
 
-const esMessages = mergeLocalizedMessages(enMessages, esCatalog);
+const esMessages = mergeLocalizedMessagesWithEmbeddedYieldBrand(englishSourceMessages, esCatalog);
 
 const frCatalog = {
   ...fr,
@@ -136,7 +132,7 @@ const frCatalog = {
   Shared: frShared,
 } satisfies LocalizedMessages<Messages>;
 
-const frMessages = mergeLocalizedMessages(enMessages, frCatalog);
+const frMessages = mergeLocalizedMessagesWithEmbeddedYieldBrand(englishSourceMessages, frCatalog);
 
 const ptCatalog = {
   ...pt,
@@ -149,7 +145,7 @@ const ptCatalog = {
   Shared: ptShared,
 } satisfies LocalizedMessages<Messages>;
 
-const ptMessages = mergeLocalizedMessages(enMessages, ptCatalog);
+const ptMessages = mergeLocalizedMessagesWithEmbeddedYieldBrand(englishSourceMessages, ptCatalog);
 
 const viCatalog = {
   ...vi,
@@ -163,7 +159,7 @@ const viCatalog = {
   Shared: viShared,
 } satisfies LocalizedMessages<Messages>;
 
-const viMessages = mergeLocalizedMessages(enMessages, viCatalog);
+const viMessages = mergeLocalizedMessagesWithEmbeddedYieldBrand(englishSourceMessages, viCatalog);
 
 export type MessageKeyFor<TValue> = TValue extends string
   ? ""
@@ -177,11 +173,11 @@ export type MessageKey = MessageKeyFor<Messages>;
 export type TranslationValues = Record<string, string | number>;
 
 const messagesByLocale: Record<AppLocale, Messages> = {
-  en: normalizeEmbeddedYieldBrand(enMessages),
-  es: normalizeEmbeddedYieldBrand(esMessages),
-  fr: normalizeEmbeddedYieldBrand(frMessages),
-  pt: normalizeEmbeddedYieldBrand(ptMessages),
-  vi: normalizeEmbeddedYieldBrand(viMessages),
+  en: englishSourceMessages,
+  es: esMessages,
+  fr: frMessages,
+  pt: ptMessages,
+  vi: viMessages,
 };
 
 export function getMessages(locale: AppLocale): Messages {
