+++ docs/contributing/earn-pluggability-playbook.md
@@ -529,7 +529,8 @@ create anything.
 
 `tsc` passes (the types are right), unit tests pass (they mock the module), lint
 is silent. **Only a browser catches it**, so finish a surfacing change by loading
-`/dashboard/markets/earn` and `/dashboard/markets/earn/deposit` for real.
+`/dashboard/markets/embedded-yield` and
+`/dashboard/markets/embedded-yield/button-builder` for real.
 
 ### Tests
